better aiming and movement mechanics

+++ include/Vector2.h
@@ -6,8 +6,8 @@ class Vector2
 public:
 	Vector2();
 	~Vector2();
-	int x = 0;
-	int y = 0;
+	float x = 0;
+	float y = 0;
 	SDL_Rect rect;
 	void MakeVel(double vel, double angR, double *ttux, double *ttuy, double delta_time, int m = 1);
 };
